refactor(benchmarks): centralize JSON validation helpers
Extract common JSON validation functions into benchmarks/_validation.py to remove duplication across benchmark-related modules. Updated imports to use these shared helpers, improving code maintainability and consistency.
